soma-research/qwen-knn/requirements.txt
torch>=2.4
transformers>=4.45
accelerate>=0.34
faiss-cpu>=1.8
numpy
tqdm
sentencepiece
einops
